REBASE ME

+++ libraries/render-utils/src/GeometryCache.cpp
@@ -12,6 +12,7 @@
 
 #include "GeometryCache.h"
 
+#include <QtCore/qtypes.h>
 #include <qmath.h>
 #include <cmath>
 
@@ -2284,7 +2285,7 @@ public:
     SimpleProgramKey(int bitmask) : _flags(bitmask) {}
 };
 
-inline uint qHash(const SimpleProgramKey& key, uint seed) {
+inline size_t qHash(const SimpleProgramKey& key, size_t seed) {
     return qHash(key.getRaw(), seed);
 }
